added base class for credential status clients; changed parameter type for github client class constructor; renamed variables
