+++ .env.example
@@ -18,4 +18,4 @@ OIDC_ISSUER_URL=https://oidc-issuer.example.com
 CRED_STATUS_REPO_NAME=credential-status
 CRED_STATUS_REPO_OWNER=university-xyz
 CRED_STATUS_REPO_VISIBILITY=public
-GITHUB_OAUTH_TOKEN=abc
+GITHUB_API_ACCESS_TOKEN=abc
