+++ README.md
@@ -16,7 +16,7 @@ Follow these steps to properly configure a `sign-and-verify` service deployment 
 - `URL`: the url where an instance of the api service will be hosted (required)
 - `OIDC_ISSUER_URL`: OIDC issuer discovery URL (required)
 - `ISSUER_MEMBERSHIP_REGISTRY_URL`: location of registry used to confirm issuer's membership status in DCC (required)
-- \*`GITHUB_OAUTH_TOKEN`: access token for GitHub API (required - necessary for GitHub-hosted credential status/revocation management)
+- \*`GITHUB_API_ACCESS_TOKEN`: access token for GitHub API (required - necessary for GitHub-hosted credential status/revocation management)
 - \*`CRED_STATUS_REPO_NAME`: name of credential status repo (required - necessary for delegated hosting of credential status/revocation management)
 - `CRED_STATUS_REPO_OWNER`: name of source code owner (required - necessary for delegated hosting of credential status/revocation management)
 - \*`CRED_STATUS_REPO_VISIBILITY`: level of visibility of credential status/revocation management repo (optional, default: `public`)
@@ -46,7 +46,7 @@ Follow these steps to properly configure a `sign-and-verify` service deployment 
 
 \*NOTE: the `DID_SEED` included as an example is just for your reference. Do not check in production did seeds, private keys, or the equivalent.
 
-\*NOTE: Here are the steps for retrieving a valid value for `GITHUB_OAUTH_TOKEN` (Please reserve this step for after you have run `npm run setup` or `yarn setup` per the instructions below):
+\*NOTE: Here are the steps for retrieving a valid value for `GITHUB_API_ACCESS_TOKEN` (Please reserve this step for after you have run `npm run setup` or `yarn setup` per the instructions below):
 1. Retrieve access token for GitHub API
 2. Login to GitHub as an authorized member of the organization
 3. Click on your profile dropdown icon in the top-right corner of the screen
@@ -57,7 +57,7 @@ Follow these steps to properly configure a `sign-and-verify` service deployment 
 8. Select the full *repo* scope
 9. Click *Generate token*
 10. Copy the generated token
-11. \*Save the token from the previous step as the `GITHUB_OAUTH_TOKEN` environment variable the service that is hosting your organization’s instance of *sign-and-verify*
+11. \*Save the token from the previous step as the `GITHUB_API_ACCESS_TOKEN` environment variable at the hosting service that is hosting your organization’s instance of *sign-and-verify*
 
 \*NOTE: `CRED_STATUS_REPO_NAME`, the designated credential status repo, will automatically be generated after running `npm run start` or `yarn start`, per the instructions below. After this repo is autogenerated, here are the steps for enabling GitHub Pages hosting for credential status lists from this repo:
 1. Select the *Settings* tab in the credential status repo
@@ -72,7 +72,7 @@ Follow these steps to properly configure a `sign-and-verify` service deployment 
 
 \*NOTE: Here are some important points on credential status management to consider in a live deployment:
 - According to [this GitHub Pages resource](https://docs.github.com/en/pages/getting-started-with-github-pages/about-github-pages): *GitHub Pages is available in public repositories with GitHub Free and GitHub Free for organizations, and in public and private repositories with GitHub Pro, GitHub Team, GitHub Enterprise Cloud, and GitHub Enterprise Server. For more information, see "[GitHub's products](https://docs.github.com/en/get-started/learning-about-github/githubs-products)."*
-- Take note of the expiration date of `GITHUB_OAUTH_TOKEN` value generated above, so that you can renew this binding in advance. Failure to do this will result in a faulty credential status management process.
+- Take note of the expiration date of `GITHUB_API_ACCESS_TOKEN` value generated above, so that you can renew this binding in advance. Failure to do this will result in a faulty credential status management process.
 - Issuers should wait a couple of minutes after configuring GitHub Pages for credential status management before issuing credentials leveraging the hosted status list. This is because there is a delay between configuration and publication.
 - Verifiers should accept that a credential that was revoked within the last couple of minutes may register as verified due to caching.
